+++ gg.js
@@ -1,3 +1,7 @@
 // Board-Specific Variables
 const TILE_WIDTH = 64;
 const GRID_W = 10, GRID_H = 8;
+
+// Initializing Map class objects.
+const initialMap = new mapObject("IT'S ALIVE!!!", MAP0);
+const rMap1 = new mapObject("Over Yonder", MAP1);