+++ ggObject.js
@@ -1,0 +1,6 @@
+class mapObject {
+    constructor(mapName, grid) {
+        this.mapName = mapName;
+        this.grid = grid;
+    }
+}