+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # October 14, 2025
 - Erased and recreated the repository to reflect my new code, beginning from scratch.
 - Began scaffolding with the HTML, CSS, Default JS file, and the Maps JS file.
+
+# October 15, 2025
+- Added mapObject class object
+- Added Map array for 2nd map
+- Named the maps